auctions/web_app/requirements.txt
bcrypt==3.1.6
Flask==1.0.2
Flask-Injector==0.12.0
Flask-Security==3.0.0
marshmallow==3.0.0b20
Inject==3.4.0
pybuses==1.1.0
pytest==4.5.0
python-dotenv==0.10.1
rq==1.0
auctions
auctions_infrastructure
customer_relationship
foundation
db_infrastructure
